01_Load_to_Bigquery/pyproject.toml
[tool.poetry]
name = "01-load-to-bigquery"
version = "0.1.0"
description = "Load data from different sources to BigQuery"
authors = ["matheusmafraoandrade <matheusmafrabh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
google-cloud-bigquery = "^3.25.0"
ipykernel = "^6.29.5"
python-dotenv = "^1.0.1"
google-cloud-storage = "^2.17.0"
dlt = {extras = ["bigquery"], version = "^0.5.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
